backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.1.0
firebase-admin==6.2.0
python-dotenv==1.0.0
django-rest-framework==0.1.0
